Updated tables

+++ Database/Tables/IncomeType.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[IncomeType] (
-    [IncomeTypeId] BIGINT        NOT NULL,
+    [IncomeTypeId] INT        NOT NULL,
     [IncomeType]   NVARCHAR (50) NULL,
     [IsActive]     BIT           NULL,
     CONSTRAINT [PK_IncomeType] PRIMARY KEY CLUSTERED ([IncomeTypeId] ASC)
